requirements.txt
Flask>=3.0
python-dotenv>=1.0
pandas>=2.0
matplotlib>=3.8
yfinance>=0.2.40
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
PyMySQL>=1.1
Flask-Login>=0.6
cryptography